requirements.txt
tensorflow-gpu==2.9.0
numpy==1.23.0
matplotlib==3.5.2
pillow==9.2.0
scipy==1.8.1